requirements.txt
websockets
redis
python-dotenv